refactor: union operation logic

+++ apps/nestjs-backend/src/features/calculation/batch.service.ts
@@ -357,31 +357,37 @@ export class BatchService {
     rawOps: { rawOp: IRawOp; docId: string }[]
   ) {
     const userId = this.cls.get('user.id');
-    const insertRowsData = rawOps.map(({ rawOp, docId }) => {
-      return {
-        collection: collectionId,
-        doc_type: docType,
-        doc_id: docId,
-        version: rawOp.v,
-        operation: JSON.stringify(rawOp),
-        created_by: userId,
-        created_time: new Date().toISOString(),
-      };
-    });
+    const insertRowsData = rawOps
+      .filter(({ rawOp }) => !('del' in rawOp && rawOp.del))
+      .map(({ rawOp, docId }) => {
+        return {
+          collection: collectionId,
+          doc_type: docType,
+          doc_id: docId,
+          version: rawOp.v,
+          operation: JSON.stringify(rawOp),
+          created_by: userId,
+          created_time: new Date().toISOString(),
+        };
+      });
 
     // delete history op when doc is deleted
     const deleteIds = rawOps
       .filter(({ rawOp }) => 'del' in rawOp && rawOp.del)
       .map(({ docId }) => docId);
 
-    const deleteOpsSql = this.knex('ops')
-      .where('collection', collectionId)
-      .whereIn('doc_id', deleteIds)
-      .delete()
-      .toQuery();
-    await this.prismaService.txClient().$executeRawUnsafe(deleteOpsSql);
+    if (deleteIds.length) {
+      const deleteOpsSql = this.knex('ops')
+        .where('collection', collectionId)
+        .whereIn('doc_id', deleteIds)
+        .delete()
+        .toQuery();
+      await this.prismaService.txClient().$executeRawUnsafe(deleteOpsSql);
+    }
 
-    const batchInsertOpsSql = this.dbProvider.batchInsertSql('ops', insertRowsData);
-    return this.prismaService.txClient().$executeRawUnsafe(batchInsertOpsSql);
+    if (insertRowsData.length) {
+      const batchInsertOpsSql = this.dbProvider.batchInsertSql('ops', insertRowsData);
+      await this.prismaService.txClient().$executeRawUnsafe(batchInsertOpsSql);
+    }
   }
 }
